Refactoring to match DL4J layer heirarchy changes
