+++ arbiter-deeplearning4j/src/main/java/org/deeplearning4j/arbiter/layers/LocalResponseNormalizationLayerSpace.java
@@ -75,7 +75,7 @@ public class LocalResponseNormalizationLayerSpace extends LayerSpace<LocalRespon
     }
 
 
-    public static class Builder extends LayerSpace.Builder<Builder> {
+    public static class Builder extends BaseLayerSpace.Builder<Builder> {
 
         private ParameterSpace<Double> n;
         private ParameterSpace<Double> k;
