+++ arbiter-deeplearning4j/src/main/java/org/deeplearning4j/arbiter/layers/VariationalAutoencoderLayerSpace.java
@@ -164,5 +164,6 @@ public class VariationalAutoencoderLayerSpace extends BasePretrainNetworkLayerSp
         public <E extends LayerSpace> E build() {
             return (E)new VariationalAutoencoderLayerSpace(this);
         }
+
     }
 }
